requirements.txt
gradio>=4.0.0
opencv-python-headless>=4.8.0
ultralytics>=8.0.0
pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
